Add a timestamp to the unique tag.
The tag was introduced to prevent older images from being
garbage collected from the registry once a newer image is built.

If a new image is built as a consequence of re-running the github action
then the GITHUB_SHA (and so the unique_tag) stays the same and the previous
image with this unique_tag is garbage collected once a new one is pushed
to the registry.

Adding a timestamp to the unique_tag prevents this from happening.